refactor: result page, and move page-only components in each page folder
