+++ src/pages/result/bottom-buttons/index.ts
@@ -1,0 +1,3 @@
+import { BottomButtons } from './BottomButtons';
+
+export default BottomButtons;
